[yangqian] add swagger2 and simply colean code

+++ src/main/java/com/thoughtworks/homework/service/PostService.java
@@ -1,8 +1,9 @@
 package com.thoughtworks.homework.service;
 
 import com.thoughtworks.homework.dto.PostResponse;
-import com.thoughtworks.homework.entity.Post;
-import com.thoughtworks.homework.entity.User;
+import com.thoughtworks.homework.entity.Posts;
+import com.thoughtworks.homework.entity.Users;
+import com.thoughtworks.homework.exception.AuthorizationException;
 import com.thoughtworks.homework.exception.BasePostException;
 import com.thoughtworks.homework.repository.PostRespository;
 import com.thoughtworks.homework.repository.UserRepository;
@@ -10,6 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.Optional;
 
 @Service
@@ -21,70 +24,68 @@ public class PostService {
     @Autowired
     private PostRespository postRespository;
 
-    private User getUserInfo() {
+    private Users getUserInfo() {
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        Optional<User> user = userRepository.findUserByEmail((String) principal);
+        Optional<Users> user = userRepository.findUserByEmail((String) principal);
         return user.get();
     }
 
-    public PostResponse<Iterable<Post>> getAllPosts(){
-        PostResponse<Iterable<Post>> postResponse = new PostResponse<>();
+    private PostResponse<Posts> generatePostRes(int code, String message, Posts data){
+        PostResponse<Posts> postPostResponse = new PostResponse<>();
+        postPostResponse.setCode(code);
+        postPostResponse.setMessage(message);
+        postPostResponse.setData(data);
+        return postPostResponse;
+    }
+
+    public PostResponse<Iterable<Posts>> getAllPosts(){
+        PostResponse<Iterable<Posts>> postResponse = new PostResponse<>();
         postResponse.setCode(200);
         postResponse.setMessage("文章数据获取成功！");
         postResponse.setData(postRespository.findAll());
         return postResponse;
     }
 
-    public PostResponse<Post> newPost(Post post) {
-        User user = getUserInfo();
-        Post p = new Post(post.getTitle(),post.getContent(),user);
+    public PostResponse<Posts> newPost(Posts posts) {
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+        String strStartTime = sdf.format(new Date());
+        Users users = getUserInfo();
+        Posts p = new Posts(posts.getTitle(), posts.getContent(),strStartTime, users);
         postRespository.save(p);
-        PostResponse<Post> postPostResponse = new PostResponse<>();
-        postPostResponse.setCode(200);
-        postPostResponse.setMessage("文章发表成功！");
-        postPostResponse.setData(p);
-        return postPostResponse;
+        return generatePostRes(200,"文章发表成功！",p);
     }
 
-    public PostResponse<Post> findPost(Integer id) {
-        Optional<Post> p = postRespository.findById(id);
+    public PostResponse<Posts> findPost(Integer id) {
+        Optional<Posts> p = postRespository.findById(id);
         if (!p.isPresent()){
             throw new BasePostException("该文章不存在！");
         }
         postRespository.save(p.get());
-        PostResponse<Post> postPostResponse = new PostResponse<>();
-        postPostResponse.setCode(200);
-        postPostResponse.setMessage("文章查找成功！");
-        postPostResponse.setData(p.get());
-        return postPostResponse;
+        return generatePostRes(200,"文章查找成功！",p.get());
     }
 
-    public PostResponse<Post> updatePost(Post post) {
-        Optional<Post> p = postRespository.findById(post.getId());
+    public PostResponse<Posts> updatePost(Posts posts) {
+        Optional<Posts> p = postRespository.findById(posts.getId());
         if (!p.isPresent()){
             throw new BasePostException("该文章不存在！");
         }
-        p.get().setTitle(post.getTitle());
-        p.get().setContent(post.getContent());
+        p.get().setTitle(posts.getTitle());
+        p.get().setContent(posts.getContent());
         postRespository.save(p.get());
-        PostResponse<Post> postPostResponse = new PostResponse<>();
-        postPostResponse.setCode(200);
-        postPostResponse.setMessage("文章更新成功！");
-        postPostResponse.setData(p.get());
-        return postPostResponse;
+        return generatePostRes(200,"文章更新成功！",p.get());
     }
 
-    public PostResponse<Post> deletePost(Integer id) {
-        Optional<Post> p = postRespository.findById(id);
-        if (!p.isPresent()){
-            throw new BasePostException("该文章不存在！");
+    public PostResponse<Posts> deletePost(Integer id) {
+        Users u = getUserInfo();
+        if (u.getId().equals(id) || u.getRole().equals("ROLE_ADMIN")){
+            Optional<Posts> p = postRespository.findById(id);
+            if (!p.isPresent()){
+                throw new BasePostException("该文章不存在！");
+            }
+            postRespository.deleteById(id);
+            return generatePostRes(200,"文章删除成功！",p.get());
         }
-        postRespository.deleteById(id);
-        PostResponse<Post> postPostResponse = new PostResponse<>();
-        postPostResponse.setCode(200);
-        postPostResponse.setMessage("文章删除成功！");
-        postPostResponse.setData(p.get());
-        return postPostResponse;
+        throw new AuthorizationException("您没有删除此文章的权限！");
     }
 
 }
